Added Shooting to the player character weapon.
Imported sound files from 3rd party github from udemy course instructor.
Added a particle system and soundcue variable to the ShooterCharacter script.
Added a socket(BarrelSocket) to the weapon bone on the Belica character skeleton.
Added checks and activations for both soundcue and particlesystem when FireWeapon is called.
BarrelSocket (weapon socket) is used as a spawn location for the particlesystem and later bullet spawn.